Update README.md
In 123 movies, there are available movies and series on the web. 123movie is the best platform to search and watch movies. This application provides the best quality of movies on the internet. You can index the link of movies also available on the internet. The speed of streaming is very high. The world is no longer the same, where only rich people use tech gadgets or maybe the internet. Most of the people now have the fastest internet with smartphones. Life has become completely easier. After the introduction of Starlink, now even remote areas have internet access.
<a href="https://vidapk.com//">Visit VidMate 2014.com!</a>
<a href="https://vidapk.com/vidmate-2014//">Visit VidMate 2014.com!</a>

<a href="https://vidapk.com/123-movie//">Visit 123 movie.com!</a>